+++ internal/elasticsearch/index/ilm_test.go
@@ -128,6 +128,14 @@ func TestAccResourceILMRolloverConditions(t *testing.T) {
 		ProtoV6ProviderFactories: acctest.Providers,
 		Steps: []resource.TestStep{
 			{
+				SkipFunc: versionutils.CheckIfVersionIsUnsupported(index.MaxPrimaryShardDocsMinSupportedVersion),
+				Config:   testAccResourceILMCreateWithMaxPrimaryShardDocs(policyName),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr("elasticstack_elasticsearch_index_lifecycle.test_rollover", "name", policyName),
+					resource.TestCheckResourceAttr("elasticstack_elasticsearch_index_lifecycle.test_rollover", "hot.0.rollover.0.max_primary_shard_docs", "5000"),
+				),
+			},
+			{
 				SkipFunc: versionutils.CheckIfVersionIsUnsupported(index.RolloverMinConditionsMinSupportedVersion),
 				Config:   testAccResourceILMCreateWithRolloverConditions(policyName),
 				Check: resource.ComposeTestCheckFunc(
@@ -135,12 +143,13 @@ func TestAccResourceILMRolloverConditions(t *testing.T) {
 					resource.TestCheckResourceAttr("elasticstack_elasticsearch_index_lifecycle.test_rollover", "hot.0.rollover.0.max_age", "7d"),
 					resource.TestCheckResourceAttr("elasticstack_elasticsearch_index_lifecycle.test_rollover", "hot.0.rollover.0.max_docs", "10000"),
 					resource.TestCheckResourceAttr("elasticstack_elasticsearch_index_lifecycle.test_rollover", "hot.0.rollover.0.max_size", "100gb"),
+					resource.TestCheckResourceAttr("elasticstack_elasticsearch_index_lifecycle.test_rollover", "hot.0.rollover.0.max_primary_shard_docs", "5000"),
 					resource.TestCheckResourceAttr("elasticstack_elasticsearch_index_lifecycle.test_rollover", "hot.0.rollover.0.max_primary_shard_size", "50gb"),
 					resource.TestCheckResourceAttr("elasticstack_elasticsearch_index_lifecycle.test_rollover", "hot.0.rollover.0.min_age", "3d"),
 					resource.TestCheckResourceAttr("elasticstack_elasticsearch_index_lifecycle.test_rollover", "hot.0.rollover.0.min_docs", "1000"),
 					resource.TestCheckResourceAttr("elasticstack_elasticsearch_index_lifecycle.test_rollover", "hot.0.rollover.0.min_size", "50gb"),
-					resource.TestCheckResourceAttr("elasticstack_elasticsearch_index_lifecycle.test_rollover", "hot.0.rollover.0.min_primary_shard_size", "25gb"),
 					resource.TestCheckResourceAttr("elasticstack_elasticsearch_index_lifecycle.test_rollover", "hot.0.rollover.0.min_primary_shard_docs", "500"),
+					resource.TestCheckResourceAttr("elasticstack_elasticsearch_index_lifecycle.test_rollover", "hot.0.rollover.0.min_primary_shard_size", "25gb"),
 				),
 			},
 		},
@@ -273,12 +282,37 @@ resource "elasticstack_elasticsearch_index_lifecycle" "test_rollover" {
       max_age = "7d"
       max_docs = 10000
       max_size = "100gb"
+      max_primary_shard_docs = 5000
       max_primary_shard_size = "50gb"
       min_age = "3d"
       min_docs = 1000
       min_size = "50gb"
-      min_primary_shard_size = "25gb"
       min_primary_shard_docs = 500
+      min_primary_shard_size = "25gb"
+    }
+
+    readonly {}
+  }
+
+  delete {
+    delete {}
+  }
+}
+ `, name)
+}
+
+func testAccResourceILMCreateWithMaxPrimaryShardDocs(name string) string {
+	return fmt.Sprintf(`
+provider "elasticstack" {
+  elasticsearch {}
+}
+
+resource "elasticstack_elasticsearch_index_lifecycle" "test_rollover" {
+  name = "%s"
+
+  hot {
+    rollover {
+      max_primary_shard_docs = 5000
     }
 
     readonly {}